fix(pyod): support vector features; clean mypy unused ignores

+++ sdks/python/apache_beam/ml/transforms/handlers.py
@@ -336,9 +336,10 @@ class TFTProcessHandler(ProcessHandler[tft_process_handler_input_type,
       if feature_type == schema_pb2.FeatureType.FLOAT:
         transformed_types[name] = Sequence[np.float32]
       elif feature_type == schema_pb2.FeatureType.INT:
-        transformed_types[name] = Sequence[np.int64]  # type: ignore[assignment]
+        # Use built-in int for portability across numpy versions.
+        transformed_types[name] = Sequence[int]
       else:
-        transformed_types[name] = Sequence[bytes]  # type: ignore[assignment]
+        transformed_types[name] = Sequence[bytes]
     return transformed_types
 
   def expand(
